+++ src/__mocks__/next/image.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable @next/next/no-img-element */
+import React from 'react';
+
+const NextImage = ({
+  src,
+  alt,
+  className,
+}: {
+  src: string;
+  alt: string;
+  className?: string;
+}) => <img src={src} alt={alt} className={className} />;
+
+export default NextImage;