+++ src/styles/media.ts
@@ -1,9 +1,0 @@
-const media = {
-  xs: `@media (min-width: 320px)`,
-  sm: `@media (min-width: 576px)`,
-  md: `@media (min-width: 768px)`,
-  lg: `@media (min-width: 1024px)`,
-  xl: `@media (min-width: 1280px)`,
-};
-
-export default media;